+++ versions.tf
@@ -4,12 +4,12 @@
 # ---------------------------------------------------------------------------------------------------------------------
 
 terraform {
-  required_version = ">= 0.14, < 2.0"
+  required_version = "~> 1.0"
 
   required_providers {
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = ">= 3.60, < 5.0"
+      version = "~> 4.9"
     }
   }
 }